+++ src/core/index.js
@@ -1,0 +1,5 @@
+import interval from './interval'
+import range from './range'
+
+
+export default {interval, range}