+++ src/core/range.spec.js
@@ -9,22 +9,22 @@ import ChunkifyOptions from '../options'
 
 test('should require a function', t => {
   t.throws(() => {
-    chunkify.loop()
-  }, /Usage: chunkify.loop\(Function fn, Number range, \[Object options]\) - bad fn/);
+    chunkify.range()
+  }, /Usage: chunkify.range\(Function fn, Number range, \[Object options]\) - bad fn/);
   t.end()
 });
 
 test('should require a function', t => {
   t.throws(() => {
-    chunkify.loop(() => {})
-  }, /Usage: chunkify.loop\(Function fn, Number range, \[Object options]\) - bad range/);
+    chunkify.range(() => {})
+  }, /Usage: chunkify.range\(Function fn, Number range, \[Object options]\) - bad range/);
   t.end()
 });
 
 test('should deserialize options', t => {
   ChunkifyOptionsSpy((spy) => {
     let options = {};
-    chunkify.loop(sinon.spy(), 10, options);
+    chunkify.range(sinon.spy(), 10, options);
     t.ok(spy.calledWith(options));
     t.end()
   });
@@ -32,7 +32,7 @@ test('should deserialize options', t => {
 
 test('should default options to an empty object', t => {
   ChunkifyOptionsSpy((spy) => {
-    chunkify.loop(sinon.spy(), 10);
+    chunkify.range(sinon.spy(), 10);
     t.ok(spy.calledWith({}));
     t.end()
   });
@@ -40,21 +40,21 @@ test('should default options to an empty object', t => {
 
 test('should remove any start value in options', t => {
   ChunkifyOptionsSpy((spy) => {
-    chunkify.loop(sinon.spy(), 10, {start: 5});
+    chunkify.range(sinon.spy(), 10, {start: 5});
     t.ok(spy.calledWith({}));
     t.end()
   });
 });
 
 test('should return a promise', t => {
-  t.ok(chunkify.loop(sinon.spy(), 10) instanceof Promise);
+  t.ok(chunkify.range(sinon.spy(), 10) instanceof Promise);
   t.end()
 });
 
 test('should not invoke fn when range is 0', t => {
   let fn = sinon.spy();
 
-  chunkify.loop(fn, 0).then(() => {
+  chunkify.range(fn, 0).then(() => {
     t.notOk(fn.called);
     t.end();
   });
@@ -63,7 +63,7 @@ test('should not invoke fn when range is 0', t => {
 test('should invoke fn with the loop index', t => {
   let fn = sinon.spy();
 
-  chunkify.loop(fn, 3, {chunk: 3}).then(() => {
+  chunkify.range(fn, 3, {chunk: 3}).then(() => {
     t.equals(fn.callCount, 3);
     t.deepEqual(fn.getCall(0).args, [0]);
     t.deepEqual(fn.getCall(1).args, [1]);
@@ -75,7 +75,7 @@ test('should invoke fn with the loop index', t => {
 test('should invoke fn with the default scope', t => {
   let fn = sinon.spy();
 
-  chunkify.loop(fn, 3, {chunk: 3}).then(() => {
+  chunkify.range(fn, 3, {chunk: 3}).then(() => {
     t.ok(fn.alwaysCalledOn(null));
     t.end()
   });
@@ -85,7 +85,7 @@ test('should invoke fn with the provided scope', t => {
   let fn = sinon.spy();
   let scope = {};
 
-  chunkify.loop(fn, 3, {chunk: 3, scope}).then(() => {
+  chunkify.range(fn, 3, {chunk: 3, scope}).then(() => {
     t.ok(fn.alwaysCalledOn(scope));
     t.end()
   });
@@ -95,29 +95,28 @@ test('should yield for at least `delay` ms after `chunk` iterations', t => {
   let fn = sinon.spy();
 
   tick({
-    delay: 999,
+    delay: 9,
 
     before_tick() {
-      chunkify.loop(fn, 4, {chunk: 3, delay: 1000});
+      chunkify.range(fn, 4, {chunk: 3, delay: 10});
       t.equals(fn.callCount, 3);
     },
 
     after_tick() {
-      // 1000 milliseconds hasn't elapsed yet
       t.equals(fn.callCount, 3);
       t.end();
     }
   });
 });
 
-test('should start again in `delay` milliseconds after yielding', t => {
+test('should start again after `delay` milliseconds from last yielding', t => {
   let fn = sinon.spy();
 
   tick({
-    delay: 1000,
+    delay: 11,
 
     before_tick() {
-      chunkify.loop(fn, 4, {chunk: 3, delay: 1000});
+      chunkify.range(fn, 4, {chunk: 3, delay: 10});
       t.equals(fn.callCount, 3);
     },
 
@@ -132,10 +131,10 @@ test('should resolve with undefined', t => {
   let fn = sinon.spy();
 
   tick({
-    delay: 1000,
+    delay: 11,
 
     before_tick() {
-      let promise = chunkify.loop(fn, 4, {chunk: 3, delay: 1000});
+      let promise = chunkify.range(fn, 4, {chunk: 3, delay: 10});
       t.equals(fn.callCount, 3);
       return promise
     },
@@ -158,7 +157,7 @@ test('should reject the promise with rejection object and stop processing', t =>
     }
   });
 
-  chunkify.loop(fn, 3, {chunk: 3}).then(null, (rejection) => {
+  chunkify.range(fn, 3, {chunk: 3}).then(null, (rejection) => {
     t.deepEquals(rejection, {error, index: 1});
     t.equals(fn.callCount, 2);
     t.end()
@@ -169,10 +168,10 @@ test('should not yield after `chunk` iterations if processing is complete', t =>
   let fn = sinon.spy();
 
   tick({
-    delay: 2000,
+    delay: 20,
 
     before_tick() {
-      chunkify.loop(fn, 3, {chunk: 3, delay: 1000});
+      chunkify.range(fn, 3, {chunk: 3, delay: 10});
       t.equals(fn.callCount, 3)
     },