+++ src/index.js
@@ -1,11 +1,13 @@
 import 'babel/polyfill'
 import arrays from './arrays'
-import loops from './loops'
+import core from './core'
+import generator from './generator'
 
 export default {
   each: arrays.each,
   map: arrays.map,
   reduce: arrays.reduce,
-  loop: loops.loop,
-  range: loops.range
+  interval: core.interval,
+  range: core.range,
+  generator: generator.from_keywords
 }