+++ src/loops/index.js
@@ -1,5 +1,0 @@
-import loop from './loop'
-import range from './range'
-
-
-export default {loop, range}