+++ src/testutils.js
@@ -8,22 +8,17 @@ let ChunkifyOptionsSpy = (callback) => {
   ChunkifyOptions.of.restore()
 };
 
+/**
+ * Invokes `before_tick`, then advances the clock by `delay` milliseconds, then
+ * invokes `after_tick` as soon as all previously queued events have been processed.
+ */
 let tick = ({before_tick, after_tick, delay}) => {
-  let ok_tick = () => {
-    if (!_.isFunction(before_tick)) {
-      throw new Error(`Expected before_tick function, got ${before_tick}`);
-    } else if (!_.isFunction(after_tick)) {
-      throw new Error(`Expected after_tick function, got ${after_tick}`);
-    } else if (!_.isNumber(delay)) {
-      throw new Error(`Expected number delay, got ${delay}`);
-    }
-  };
-  ok_tick();
   let clock = sinon.useFakeTimers();
   let before_tick_result = before_tick();
   clock.tick(delay);
-  after_tick(before_tick_result);
   clock.restore();
+  // call `after_tick` after this stack and all its queued events have unwound
+  setTimeout(after_tick, 0, before_tick_result);
 };
 
 export default {ChunkifyOptionsSpy, tick}