+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [3.0.0] - [3.0.2] 2015-10-03
+### BREAKING CHANGES
+* `chunkify.range` renamed to `chunkify.interval`
+* `chunkify.loop` renamed to `chunkify.range`
+* asynchronous logic moved from `chunkify.interval` to `chunkify.generator`
+* exposed core API generator `chunkify.generator`
+* updated README
+
 ## [2.0.3 - 2.1.1] - 2015-09-21
 ### Added
 * `chunkify.map`